+++ COLLADAMaya/src/COLLADAMayaLightExporter.cpp
@@ -48,7 +48,7 @@ namespace COLLADAMaya
         SceneElementsList* exportNodesTree = sceneGraph->getExportNodesTree();
 
         // Export all/selected DAG nodes
-        uint length = exportNodesTree->size();
+        size_t length = exportNodesTree->size();
         for ( uint i = 0; i < length; ++i )
         {
             SceneElement* sceneElement = ( *exportNodesTree ) [i];
